+++ packages/components/fields/multiline-text-field/src/index.js
@@ -1,1 +1,3 @@
 export { default } from './multiline-text-field';
+
+export { default as version } from './version';
